+++ robot_framework/reset.py
@@ -5,7 +5,7 @@ from OpenOrchestrator.orchestrator_connection.connection import OrchestratorConn
 from itk_dev_shared_components.sap import sap_login
 
 from robot_framework import config
-from robot_framework.sub_process import structura_process
+from robot_framework.sub_process import structura_process, doc2archive_process
 
 
 def reset(orchestrator_connection: OrchestratorConnection) -> None:
@@ -31,6 +31,7 @@ def kill_all(orchestrator_connection: OrchestratorConnection) -> None:
     """Forcefully close all applications used by the robot."""
     orchestrator_connection.log_trace("Killing all applications.")
 
+    doc2archive_process.kill_doc2archive()
     structura_process.kill_structura()
     sap_login.kill_sap()
 
@@ -44,3 +45,5 @@ def open_all(orchestrator_connection: OrchestratorConnection) -> None:
 
     kmd_creds = orchestrator_connection.get_credential(config.STRUCTURA_LOGIN)
     structura_process.open_structura(kmd_creds.username, kmd_creds.password)
+
+    doc2archive_process.open_doc2archive()